sync

+++ coursera/downloaders.py
@@ -1,17 +1,28 @@
 # -*- coding: utf-8 -*-
+"""
+Module for download-related classes and functions.
+
+We currently support an internal downloader written in Python with just the
+essential functionality and four "industrial-strength" external downloaders,
+namely, aria2c, axel, curl, and wget.
+"""
 
 from __future__ import print_function
 
 import logging
 import math
 import os
-import requests
 import subprocess
 import sys
 import time
 
+import requests
+
 from six import iteritems
 
+#
+# Below are file downloaders, they are wrappers for external downloaders.
+#
 
 class Downloader(object):
     """
@@ -26,34 +37,36 @@ class Downloader(object):
       >>> d.download('http://example.com', 'save/to/this/file')
     """
 
-    def _start_download(self, url, filename):
+    def _start_download(self, url, filename, resume):
         """
         Actual method to download the given url to the given file.
         This method should be implemented by the subclass.
         """
         raise NotImplementedError("Subclasses should implement this")
 
-    def download(self, url, filename):
+    def download(self, url, filename, resume=False):
         """
         Download the given url to the given file. When the download
         is aborted by the user, the partially downloaded file is also removed.
         """
 
         try:
-            self._start_download(url, filename)
+            self._start_download(url, filename, resume)
         except KeyboardInterrupt as e:
-            logging.info(
-                'Keyboard Interrupt -- Removing partial file: %s', filename)
-            try:
-                os.remove(filename)
-            except OSError:
-                pass
+            # keep the file if resume is True
+            if not resume:
+                logging.info('Keyboard Interrupt -- Removing partial file: %s',
+                             filename)
+                try:
+                    os.remove(filename)
+                except OSError:
+                    pass
             raise e
 
 
 class ExternalDownloader(Downloader):
     """
-    Downloads files with an extrnal downloader.
+    Downloads files with an external downloader.
 
     We could possibly use python to stream files to disk,
     but this is slow compared to these external downloaders.
@@ -65,9 +78,10 @@ class ExternalDownloader(Downloader):
     # External downloader binary
     bin = None
 
-    def __init__(self, session, bin=None):
+    def __init__(self, session, bin=None, downloader_arguments=None):
         self.session = session
         self.bin = bin or self.__class__.bin
+        self.downloader_arguments = downloader_arguments or []
 
         if not self.bin:
             raise RuntimeError("No bin specified")
@@ -87,6 +101,13 @@ class ExternalDownloader(Downloader):
         if cookie_values:
             self._add_cookies(command, cookie_values)
 
+    def _enable_resume(self, command):
+        """
+        Enable resume feature
+        """
+
+        raise RuntimeError("Subclass should implement this")
+
     def _add_cookies(self, command, cookie_values):
         """
         Add the given cookie values to the command
@@ -100,9 +121,13 @@ class ExternalDownloader(Downloader):
         """
         raise NotImplementedError("Subclasses should implement this")
 
-    def _start_download(self, url, filename):
+    def _start_download(self, url, filename, resume):
         command = self._create_command(url, filename)
+        command.extend(self.downloader_arguments)
         self._prepare_cookies(command, url)
+        if resume:
+            self._enable_resume(command)
+
         logging.debug('Executing %s: %s', self.bin, command)
         try:
             subprocess.call(command)
@@ -119,6 +144,9 @@ class WgetDownloader(ExternalDownloader):
 
     bin = 'wget'
 
+    def _enable_resume(self, command):
+        command.append('-c')
+
     def _add_cookies(self, command, cookie_values):
         command.extend(['--header', "Cookie: " + cookie_values])
 
@@ -134,6 +162,9 @@ class CurlDownloader(ExternalDownloader):
 
     bin = 'curl'
 
+    def _enable_resume(self, command):
+        command.extend(['-C', '-'])
+
     def _add_cookies(self, command, cookie_values):
         command.extend(['--cookie', cookie_values])
 
@@ -148,6 +179,9 @@ class Aria2Downloader(ExternalDownloader):
     """
 
     bin = 'aria2c'
+
+    def _enable_resume(self, command):
+        command.append('-c')
 
     def _add_cookies(self, command, cookie_values):
         command.extend(['--header', "Cookie: " + cookie_values])
@@ -165,6 +199,10 @@ class AxelDownloader(ExternalDownloader):
     """
 
     bin = 'axel'
+
+    def _enable_resume(self, command):
+        logging.warn('Resume download not implemented for this '
+                     'downloader!')
 
     def _add_cookies(self, command, cookie_values):
         command.extend(['-H', "Cookie: " + cookie_values])
@@ -224,11 +262,18 @@ class DownloadProgress(object):
         self._current += bytes
         self.report_progress()
 
+    def report(self, bytes):
+        self._now = time.time()
+        self._current = bytes
+        self.report_progress()
+
     def calc_percent(self):
         if self._total is None:
             return '--%'
+        if self._total == 0:
+            return '100% done'
         percentage = int(float(self._current) / float(self._total) * 100.0)
-        done = int(percentage/2)
+        done = int(percentage / 2)
         return '[{0: <50}] {1}%'.format(done * '#', percentage)
 
     def calc_speed(self):
@@ -264,47 +309,78 @@ class NativeDownloader(Downloader):
     def __init__(self, session):
         self.session = session
 
-    def _start_download(self, url, filename):
-        logging.info('Downloading %s -> %s', url, filename)
+    def _start_download(self, url, filename, resume=False):
+        # resume has no meaning if the file doesn't exists!
+        resume = resume and os.path.exists(filename)
 
+        headers = {}
+        filesize = None
+        if resume:
+            filesize = os.path.getsize(filename)
+            headers['Range'] = 'bytes={}-'.format(filesize)
+            logging.info('Resume downloading %s -> %s', url, filename)
+        else:
+            logging.info('Downloading %s -> %s', url, filename)
+
+        max_attempts = 3
         attempts_count = 0
         error_msg = ''
-        while attempts_count < 5:
-            r = self.session.get(url, stream=True)
+        while attempts_count < max_attempts:
+            r = self.session.get(url, stream=True, headers=headers)
 
-            if r.status_code is not 200:
-                logging.warn(
-                    'Probably the file is missing from the AWS repository...'
-                    ' waiting.')
-
-                if r.reason:
-                    error_msg = r.reason + ' ' + str(r.status_code)
+            if r.status_code != 200:
+                # because in resume state we are downloading only a
+                # portion of requested file, server may return
+                # following HTTP codes:
+                # 206: Partial Content
+                # 416: Requested Range Not Satisfiable
+                # which are OK for us.
+                if resume and r.status_code == 206:
+                    pass
+                elif resume and r.status_code == 416:
+                    logging.info('%s already downloaded', filename)
+                    r.close()
+                    return True
                 else:
-                    error_msg = 'HTTP Error ' + str(r.status_code)
+                    print('%s %s %s' % (r.status_code, url, filesize))
+                    logging.warn('Probably the file is missing from the AWS '
+                                 'repository...  waiting.')
 
-                wait_interval = 2 ** (attempts_count + 1)
-                msg = 'Error downloading, will retry in {0} seconds ...'
-                print(msg.format(wait_interval))
-                time.sleep(wait_interval)
-                attempts_count += 1
-                continue
+                    if r.reason:
+                        error_msg = r.reason + ' ' + str(r.status_code)
+                    else:
+                        error_msg = 'HTTP Error ' + str(r.status_code)
+
+                    wait_interval = 2 ** (attempts_count + 1)
+                    msg = 'Error downloading, will retry in {0} seconds ...'
+                    print(msg.format(wait_interval))
+                    time.sleep(wait_interval)
+                    attempts_count += 1
+                    continue
+
+            if resume and r.status_code == 200:
+                # if the server returns HTTP code 200 while we are in
+                # resume mode, it means that the server does not support
+                # partial downloads.
+                resume = False
 
             content_length = r.headers.get('content-length')
-            progress = DownloadProgress(content_length)
             chunk_sz = 1048576
-            with open(filename, 'wb') as f:
-                progress.start()
-                while True:
-                    data = r.raw.read(chunk_sz)
-                    if not data:
-                        progress.stop()
-                        break
-                    progress.read(len(data))
-                    f.write(data)
+            progress = DownloadProgress(content_length)
+            progress.start()
+            f = open(filename, 'ab') if resume else open(filename, 'wb')
+            while True:
+                data = r.raw.read(chunk_sz, decode_content=True)
+                if not data:
+                    progress.stop()
+                    break
+                progress.report(r.raw.tell())
+                f.write(data)
+            f.close()
             r.close()
             return True
 
-        if attempts_count == 5:
+        if attempts_count == max_attempts:
             logging.warn('Skipping, can\'t download file ...')
             logging.error(error_msg)
             return False
@@ -324,6 +400,7 @@ def get_downloader(session, class_name, args):
 
     for bin, class_ in iteritems(external):
         if getattr(args, bin):
-            return class_(session, bin=getattr(args, bin))
+            return class_(session, bin=getattr(args, bin),
+                          downloader_arguments=args.downloader_arguments)
 
     return NativeDownloader(session)
